+++ packages/react/babel.config.js
@@ -4,7 +4,11 @@ module.exports = {
       "@babel/preset-env",
       { targets: "defaults, not ie 11, not ie_mob 11", modules: "commonjs" },
     ],
-    ["@babel/preset-react", { runtime: "automatic" }],
+    // Use the classic runtime until we drop support for React < 18.0
+    // because in older versions they don't have "exports" in package.json
+    // and it does not work well with ESM.
+    // See configs/tsconfig.base.json too
+    ["@babel/preset-react"],
     ["@babel/preset-typescript", { allowDeclareFields: true }],
   ],
   plugins: [