+++ packages/react/src/index.test.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { cleanup, render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { describe, expect, it } from "@jest/globals";