+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fix "Cannot find module react/jsx-runtime" error with ESM (dist/index.mjs).
+
 ## 0.1.0
 
 Initial release.